creat SortingNetwork
